Add KitchenSink.Algorithms.binary_search & tests (#62)
* Add KitchenSink.Algorithms.binary_search & tests

(Mike & Marius) and Craig were working separately on implementations of
a binary search for use in the app. This is an attempt to unify those
efforts.

* Update dev dependencies, add CHANGELOG.md, bump to 1.1.0

Minor version bump as we're adding something to the interface.

(see CHANGELOG.md ;-)

* Rename the interval version of binary_search (PR change)

As the binary search and interval binary search have very different
behaviours which are triggered by the type of the first argument, the
interval one has been renamed to `binary_search_interval` to avoid
"accidental" misuse.

`binary_search_interval` example is more in line with the real use case
for @kc.

Make binary_interval_search and binary_search arguments more consistent

  start, finish, fit, target, ...

Fit functions now take 2 args, the current position and the target
value.